Move CorsNotification into individual pages
