+++ v3/src/js/views/AppShell.jsx
@@ -6,30 +6,20 @@ import type { ModuleCode, Semester } from 'types/modules';
 import type { Mode } from 'types/settings';
 
 import React, { Component } from 'react';
-import { NavLink, withRouter, type ContextRouter } from 'react-router-dom';
+import { NavLink, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import NUSModerator from 'nusmoderator';
-import qs from 'query-string';
 import classnames from 'classnames';
 
-import config from 'config';
 import { fetchModuleList, fetchModule } from 'actions/moduleBank';
 import { noBreak } from 'utils/react';
-import { roundStart } from 'utils/cors';
 import ModulesSelect from 'views/components/ModulesSelect';
 import Footer from 'views/layout/Footer';
 import Navtabs from 'views/layout/Navtabs';
 import { DARK_MODE } from 'types/settings';
 import LoadingSpinner from './components/LoadingSpinner';
-import CorsNotification from './components/cors-info/CorsNotification';
-
-// Cache a current date object to stop CorsNotification from re-rendering - if this was in
-// render(), a new Date object is created, forcing re-render.
-const NOW = new Date();
 
 type Props = {
-  ...ContextRouter,
-
   children: Node,
   moduleList: ModuleList,
   moduleSelectList: ModuleSelectList,
@@ -72,9 +62,9 @@ function setMode(mode: Mode) {
 
   if (mode === DARK_MODE) {
     document.body.classList.add('mode-dark');
-    return;
+  } else {
+    document.body.classList.remove('mode-dark');
   }
-  document.body.classList.remove('mode-dark');
 }
 
 export class AppShell extends Component<Props> {
@@ -94,18 +84,6 @@ export class AppShell extends Component<Props> {
 
   componentWillUpdate(nextProps: Props) {
     setMode(nextProps.mode);
-  }
-
-  currentTime() {
-    const debugRound = qs.parse(this.props.location.search).round;
-
-    // For manual testing - add ?round=1A (or other round names) to trigger the notification
-    if (debugRound) {
-      const round = config.corsSchedule.find(r => r.round === debugRound);
-      if (round) return roundStart(round);
-    }
-
-    return NOW;
   }
 
   render() {
@@ -133,8 +111,6 @@ export class AppShell extends Component<Props> {
         <div className="main-container">
           <Navtabs />
 
-          <CorsNotification time={this.currentTime()} />
-
           <main className={classnames('main-content', `theme-${this.props.theme}`)}>
             {isModuleListReady ? this.props.children : <LoadingSpinner />}
           </main>
@@ -155,6 +131,10 @@ const mapStateToProps = state => ({
   activeSemester: state.app.activeSemester,
 });
 
+// withRouter here is used to ensure re-render when routes change, since
+// connect implements shouldComponentUpdate based purely on props. If it
+// is removed, connect not detect prop changes when route is changed and
+// thus the pages are not re-rendered
 export default withRouter(
   connect(mapStateToProps, {
     fetchModuleList,
