+++ v3/src/js/views/components/cors-info/CorsNotification.jsx
@@ -2,31 +2,40 @@
 
 import React, { PureComponent } from 'react';
 import { capitalize } from 'lodash';
+import { withRouter, type ContextRouter } from 'react-router-dom';
+import qs from 'query-string';
 
 import config, { type CorsRound } from 'config';
-import { roundEnd } from 'utils/cors';
+import { roundStart, roundEnd } from 'utils/cors';
 
 import styles from './CorsNotification.scss';
 
-type Props = {
-  time: Date,
-};
+type Props = ContextRouter;
 
-export default class CorsNotification extends PureComponent<Props> {
-  static defaultProps = {
-    time: new Date(),
-  };
+const NOW = new Date();
+
+export class CorsNotificationComponent extends PureComponent<Props> {
+  currentTime() {
+    const debugRound = qs.parse(this.props.location.search).round;
+
+    // For manual testing - add ?round=1A (or other round names) to trigger the notification
+    if (debugRound) {
+      const round = config.corsSchedule.find(r => r.round === debugRound);
+      if (round) return roundStart(round);
+    }
+
+    return NOW;
+  }
 
   currentCorsRound() {
-    return config.corsSchedule.find(round => roundEnd(round) > this.props.time);
+    return config.corsSchedule.find(round => roundEnd(round) > this.currentTime());
   }
 
   currentPeriod(round: CorsRound) {
-    return round.periods.find(period => period.endDate > this.props.time);
+    return round.periods.find(period => period.endDate > this.currentTime());
   }
 
   render() {
-    const { time } = this.props;
     const round = this.currentCorsRound();
 
     // CORS bidding is over - don't show anything
@@ -34,7 +43,7 @@ export default class CorsNotification extends PureComponent<Props> {
     const period = this.currentPeriod(round);
     if (!period) return null; // Shouldn't happen
 
-    const isRoundOpen = time >= period.startDate;
+    const isRoundOpen = this.currentTime() >= period.startDate;
     return (
       <a href="https://myaces.nus.edu.sg/cors/StudentLogin" target="_blank" rel="noopener noreferrer">
         <div className={styles.wrapper}>
@@ -46,3 +55,6 @@ export default class CorsNotification extends PureComponent<Props> {
     );
   }
 }
+
+const CorsNotification = withRouter(CorsNotificationComponent);
+export default CorsNotification;
