+++ v3/src/js/views/timetable/TimetableContent.jsx
@@ -41,6 +41,7 @@ import {
   findExamClashes,
 } from 'utils/timetables';
 import ModulesSelect from 'views/components/ModulesSelect';
+import CorsNotification from 'views/components/cors-info/CorsNotification';
 
 import styles from './TimetableContent.scss';
 import Timetable from './Timetable';
@@ -204,9 +205,13 @@ class TimetableContainer extends Component<Props> {
         <Helmet>
           <title>Timetable - {config.brandName}</title>
         </Helmet>
+
+        <CorsNotification />
+
         <div>
           {this.props.header}
         </div>
+
         <div className="row">
           <div
             className={classnames({
